requirements.txt
# Azure AI Services
azure-cognitiveservices-speech==1.46.0
openai==1.12.0
azure-core==1.35.0

# Streamlit
streamlit==1.31.0
audio-recorder-streamlit==0.0.10

# Audio processing
pyaudio==0.2.14
pydub==0.25.1
soundfile==0.12.1
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# RAG System Dependencies
langchain==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.5
chromadb==0.4.22
faiss-cpu==1.7.4
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2

# Additional dependencies for deployment
gunicorn==21.2.0
